Hid the bottom navigation bar from the Upload Item Screen
The user is only on this screen for a very short time (while they add the item details) so no need to display the navigation bar. They are on this screen with a purpose and once they hit Save, they will navigate away. If they no longer wish to upload an item, they can hit the back button or click cancel and they will go back to the Closet Screen