Updated the Display Item Screen to conditionally show icons
When the user navigates to this screen after choosing the Play It Safe Feature, the only icon available is the regenerate icon (no need for the favorites icon as the outfit is already pulled from favorites). When navigating to this screen from the Surprise Me Option, we show the thumbs-down, favorite, regenerate and thumbs-up icons. The thumbs up/down icons will help train the AI on the user preferences.