Reverted back to the original layout for display outfit
Using a column to display the items on top of each other requires scrolling to view all the items properly and causes the buttons for surprise me and play it safe to be hidden from view until the user scrolls down. This is not ok-- the user should see the regenerate, add to vavorites and feedback buttons as soon as they land on the screen w/o having to scroll to find them.